+++ lib/login.dart
@@ -1,0 +1,26 @@
+import 'package:cracker_app/markers.dart';
+import 'package:flutter/material.dart';
+
+class Login extends StatelessWidget {
+  final loginAction;
+  final String loginError;
+
+  const Login(this.loginAction, this.loginError);
+
+  @override
+  Widget build(BuildContext context) {
+    return Column(
+      mainAxisAlignment: MainAxisAlignment.center,
+      children: <Widget>[
+        RaisedButton(
+          onPressed: () {
+            loginAction();
+          },
+          child: Text('Login now!'),
+        ),
+        Text(loginError ?? ''),
+        Markers()
+      ],
+    );
+  }
+}
