+++ lib/main.dart
@@ -1,40 +1,8 @@
-import 'dart:convert';
-
+import 'package:cracker_app/auth.dart';
 import 'package:flutter/material.dart';
-import 'package:flutter_appauth/flutter_appauth.dart';
-import 'package:flutter_secure_storage/flutter_secure_storage.dart';
 import 'package:graphql_flutter/graphql_flutter.dart';
-import 'package:http/http.dart' as http;
-
-final FlutterAppAuth appAuth = FlutterAppAuth();
-final FlutterSecureStorage secureStorage = const FlutterSecureStorage();
-
-const AUTH0_DOMAIN = 'fulbert.eu.auth0.com';
-const AUTH0_CLIENT_ID = 'KZZWiCz61h8DNxs1DSYxPM7xfxkLDEKF';
-
-const AUTH0_REDIRECT_URI = 'com.cracker.app://login-callback';
-const AUTH0_ISSUER = 'https://$AUTH0_DOMAIN';
 
 void main() => runApp(CrackerApp());
-
-String readRepositories = """
-    {
-    markers {
-      _id
-      latitude
-      longitude
-      imageFilename
-      polish {
-        name
-        description
-      }
-      english {
-        name
-        description
-      }
-    }
-  }
-""";
 
 class CrackerApp extends StatelessWidget {
   Widget build(BuildContext context) {
@@ -48,255 +16,6 @@ class CrackerApp extends StatelessWidget {
         title: 'Cracker app',
         home: Auth0App(),
       ),
-    );
-  }
-}
-
-class Auth0App extends StatefulWidget {
-  @override
-  _Auth0AppState createState() => _Auth0AppState();
-}
-
-class _Auth0AppState extends State<Auth0App> {
-  bool isBusy = false;
-  bool isLoggedIn = false;
-  String errorMessage;
-  String name;
-  String picture;
-
-  Map<String, dynamic> parseIdToken(String idToken) {
-    final parts = idToken.split(r'.');
-    assert(parts.length == 3);
-
-    return jsonDecode(
-        utf8.decode(base64Url.decode(base64Url.normalize(parts[1]))));
-  }
-
-  Future<Map<String, dynamic>> getUserDetails(String accessToken) async {
-    final url = 'https://$AUTH0_DOMAIN/userinfo';
-    final response = await http.get(
-      url,
-      headers: {'Authorization': 'Bearer $accessToken'},
-    );
-
-    if (response.statusCode == 200) {
-      return jsonDecode(response.body);
-    } else {
-      throw Exception('Failed to get user details');
-    }
-  }
-
-  void logoutAction() async {
-    await secureStorage.delete(key: 'refresh_token');
-    setState(() {
-      isLoggedIn = false;
-      isBusy = false;
-    });
-  }
-
-  Future<void> loginAction() async {
-    setState(() {
-      isBusy = true;
-      errorMessage = '';
-    });
-
-    try {
-      final AuthorizationTokenResponse result =
-          await appAuth.authorizeAndExchangeCode(
-        AuthorizationTokenRequest(
-          AUTH0_CLIENT_ID,
-          AUTH0_REDIRECT_URI,
-          issuer: 'https://$AUTH0_DOMAIN',
-          scopes: ['openid', 'profile', 'offline_access'],
-          // promptValues: ['login']
-        ),
-      );
-
-      final idToken = parseIdToken(result.idToken);
-      final profile = await getUserDetails(result.accessToken);
-
-      await secureStorage.write(
-          key: 'refresh_token', value: result.refreshToken);
-
-      setState(() {
-        isBusy = false;
-        isLoggedIn = true;
-        name = idToken['name'];
-        picture = profile['picture'];
-      });
-    } catch (e, s) {
-      print('login error: $e - stack: $s');
-
-      setState(() {
-        isBusy = false;
-        isLoggedIn = false;
-        errorMessage = e.toString();
-      });
-    }
-  }
-
-  @override
-  void initState() {
-    initAction();
-    super.initState();
-  }
-
-  void initAction() async {
-    final storedRefreshToken = await secureStorage.read(key: 'refresh_token');
-    if (storedRefreshToken == null) return;
-
-    setState(() {
-      isBusy = true;
-    });
-
-    try {
-      final response = await appAuth.token(TokenRequest(
-        AUTH0_CLIENT_ID,
-        AUTH0_REDIRECT_URI,
-        issuer: AUTH0_ISSUER,
-        refreshToken: storedRefreshToken,
-      ));
-
-      final idToken = parseIdToken(response.idToken);
-      final profile = await getUserDetails(response.accessToken);
-
-      secureStorage.write(key: 'refresh_token', value: response.refreshToken);
-
-      setState(() {
-        isBusy = false;
-        isLoggedIn = true;
-        name = idToken['name'];
-        picture = profile['picture'];
-      });
-    } catch (e, s) {
-      print('error on refresh token: $e - stack: $s');
-      logoutAction();
-    }
-  }
-
-  @override
-  Widget build(BuildContext context) {
-    return MaterialApp(
-      title: 'Auth0 Demo',
-      home: Scaffold(
-        appBar: AppBar(
-          title: Text('Auth0 Demo'),
-        ),
-        body: Center(
-          child: isBusy
-              ? CircularProgressIndicator()
-              : isLoggedIn
-                  ? Profile(logoutAction, name, picture)
-                  : Login(loginAction, errorMessage),
-        ),
-      ),
-    );
-  }
-}
-
-class Profile extends StatelessWidget {
-  final logoutAction;
-  final String name;
-  final String picture;
-
-  Profile(this.logoutAction, this.name, this.picture);
-
-  @override
-  Widget build(BuildContext context) {
-    return Column(
-      mainAxisAlignment: MainAxisAlignment.center,
-      children: <Widget>[
-        Container(
-          width: 150,
-          height: 150,
-          decoration: BoxDecoration(
-            border: Border.all(color: Colors.blue, width: 4.0),
-            shape: BoxShape.circle,
-            image: DecorationImage(
-              fit: BoxFit.fill,
-              image: NetworkImage(picture ?? ''),
-            ),
-          ),
-        ),
-        SizedBox(height: 24.0),
-        Text('Name: $name'),
-        SizedBox(height: 48.0),
-        RaisedButton(
-          onPressed: () {
-            logoutAction();
-          },
-          child: Text('Logout'),
-        ),
-      ],
-    );
-  }
-}
-
-class Login extends StatelessWidget {
-  final loginAction;
-  final String loginError;
-
-  const Login(this.loginAction, this.loginError);
-
-  @override
-  Widget build(BuildContext context) {
-    return Column(
-      mainAxisAlignment: MainAxisAlignment.center,
-      children: <Widget>[
-        RaisedButton(
-          onPressed: () {
-            loginAction();
-          },
-          child: Text('Login'),
-        ),
-        Text(loginError ?? ''),
-        Markers()
-      ],
-    );
-  }
-}
-
-class Markers extends StatelessWidget {
-  final _biggerFont = TextStyle(fontSize: 18.0);
-
-  Widget _buildRow(dynamic marker) {
-    return ListTile(
-      title: Text(
-        marker['english']['name'],
-        style: _biggerFont,
-      ),
-    );
-  }
-
-  @override
-  Widget build(BuildContext context) {
-    return Query(
-      options: QueryOptions(
-        documentNode: gql(readRepositories),
-      ),
-      builder: (QueryResult result,
-          {VoidCallback refetch, FetchMore fetchMore}) {
-        if (result.hasException) {
-          return Text(result.exception.toString());
-        }
-
-        if (result.loading) {
-          return Text('Loading');
-        }
-
-        List markers = result.data['markers'];
-
-        return Text(markers[0]['english']['name']);
-
-        /*
-        return ListView.builder(
-            itemCount: markers.length,
-            itemBuilder: (context, index) {
-              final marker = markers[index];
-              return _buildRow(marker);
-            });
-            */
-      },
     );
   }
 }
